gravando jogos avaliados em arquivo json

+++ src/main/kotlin/br/com/teste/alugames/model/Jogo.kt
@@ -3,6 +3,7 @@ package org.example.br.com.teste.alugames.model
 import br.com.teste.alugames.model.Recomendavel
 import com.google.gson.annotations.Expose
 import formatoComDuasCasasDecimais
+import java.math.BigDecimal
 
 data class Jogo(
     @Expose var titulo: String,
@@ -11,14 +12,14 @@ data class Jogo(
 
     var descricao: String? = null
 
-    var preco = 0.0
+    var preco = BigDecimal(0.0)
 
     private val listaNotas = mutableListOf<Int>()
     override val media: Double
         get() = listaNotas.average().formatoComDuasCasasDecimais()
 
 
-    constructor(titulo: String, capa: String, preco: Double, descricao: String) :
+    constructor(titulo: String, capa: String, preco: BigDecimal, descricao: String) :
             this(titulo, capa) {
         this.preco = preco
         this.descricao = descricao
